requirements.txt
streamlit
moviepy
openai-whisper
ffmpeg-python
pydub
opencv-python
